backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM Integration
google-generativeai==0.3.2
ollama==0.1.6

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Financial Data APIs
yfinance==0.2.35

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.25.2

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
